Major refactoring of pursuit and subset learners
