+++ adam/experiment/log_experiment.py
@@ -19,9 +19,9 @@ from adam.experiment import Experiment, execute_experiment
 from adam.experiment.observer import LearningProgressHtmlLogger, CandidateAccuracyObserver
 from adam.learner import LanguageLearner
 from adam.learner.object_recognizer import ObjectRecognizer
-from adam.learner.preposition_subset import PrepositionSubsetLearner
+from adam.learner.prepositions import SubsetPrepositionLearner
 from adam.learner.pursuit import HypothesisLogger
-from adam.learner.pursuit.object import ObjectPursuitLearner
+from adam.learner.objects import ObjectPursuitLearner
 from adam.ontology.phase1_ontology import GAILA_PHASE_1_ONTOLOGY
 from adam.perception.high_level_semantics_situation_to_developmental_primitive_perception import (
     GAILA_M6_PERCEPTION_GENERATOR,
@@ -73,7 +73,7 @@ def learner_factory_from_params(
             params.namespace("pursuit"), graph_logger=graph_logger
         )
     elif learner_type == "preposition-subset":
-        return lambda: PrepositionSubsetLearner(
+        return lambda: SubsetPrepositionLearner(
             graph_logger=graph_logger,
             # Eval hack! This is specific to the M6 ontology
             object_recognizer=ObjectRecognizer.for_ontology_types(
