+++ tests/learner/pursuit_object_learner_test.py
@@ -7,8 +7,7 @@ from adam.curriculum.phase1_curriculum import PHASE1_CHOOSER, phase1_instances
 from adam.curriculum.pursuit_curriculum import make_simple_pursuit_curriculum
 from adam.language_specific.english.english_language_generator import IGNORE_COLORS
 from adam.learner import LearningExample
-from adam.learner.pursuit.object import ObjectPursuitLearner
-from adam.learner.subset import SubsetObjectLearner
+from adam.learner.objects import ObjectPursuitLearner, SubsetObjectLearner
 from adam.ontology import OntologyNode
 from adam.ontology.phase1_ontology import (
     BALL,
@@ -54,9 +53,7 @@ def run_subset_learner_for_object(
         ),
     )
 
-    learner = SubsetObjectLearner(  # type: ignore
-        debug_callback=debug_callback, ontology=GAILA_PHASE_1_ONTOLOGY
-    )
+    learner = SubsetObjectLearner(ontology=GAILA_PHASE_1_ONTOLOGY)
     for training_stage in [obj_curriculum]:
         for (
             _,
